# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.FileDTO:
    properties:
      contentType:
        type: string
      createdAt:
        type: string
      filename:
        type: string
      id:
        type: string
      offerId:
        type: string
      size:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@straye.io
    name: API Support
  description: CRM Relation API for customer, project, and sales pipeline management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Straye Relation API
  version: "1.0"
paths:
...
  /files/{id}:
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.FileDTO'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get file metadata
      tags:
      - Files
  /files/{id}/download:
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Download file
      tags:
      - Files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Offer ID to attach file to
        in: formData
        name: offerId
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.FileDTO'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Upload file
      tags:
      - Files
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for system operations
    in: header
    name: x-api-key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"